nfdiv-2350 move application submitted to app1 sequence
