+++ src/main/steps/applicant1/application-submitted/content.ts
@@ -1,13 +1,13 @@
 import config from 'config';
 import dayjs from 'dayjs';
 
-import { Applicant2Represented, DocumentType, State, YesOrNo } from '../../app/case/definition';
-import { TranslationFn } from '../../app/controller/GetController';
-import { getProgressBarContent } from '../applicant1/hub-page/progressBarLabels';
-import { isCountryUk } from '../applicant1Sequence';
-import type { CommonContent } from '../common/common.content';
-import { formattedCaseId } from '../common/content.utils';
-import { currentStateFn } from '../state-sequence';
+import { Applicant2Represented, DocumentType, State, YesOrNo } from '../../../app/case/definition';
+import { TranslationFn } from '../../../app/controller/GetController';
+import { isCountryUk } from '../../applicant1Sequence';
+import type { CommonContent } from '../../common/common.content';
+import { formattedCaseId } from '../../common/content.utils';
+import { currentStateFn } from '../../state-sequence';
+import { getProgressBarContent } from '../hub-page/progressBarLabels';
 
 const en = ({
   isDivorce,
