+++ src/main/steps/applicant1/application-submitted/progress-bar.njk
