+++ src/main/steps/applicant1Sequence.ts
@@ -413,6 +413,10 @@ export const applicant1PostSubmissionSequence: Step[] = [
     getNextStep: () => APPLICATION_SUBMITTED,
   },
   {
+    url: APPLICATION_SUBMITTED,
+    getNextStep: () => HOME_URL,
+  },
+  {
     url: HUB_PAGE,
     getNextStep: () => HOME_URL,
   },
