--- src/main/steps/application-submitted/get.ts
@@ -1,9 +1,0 @@
-import { GetController } from '../../app/controller/GetController';
-
-import { generateContent } from './content';
-
-export class ApplicationSubmittedGetController extends GetController {
-  constructor() {
-    super(__dirname + '/template.njk', generateContent);
-  }
-}
